refactor: Controller의 복잡한 while문을 최적화 및 Command 패키지 생성
