+++ src/main/java/command/StartCommand.java
@@ -1,4 +1,4 @@
-package controller;
+package command;
 
 import java.util.Arrays;
